refactor: moved into one module and multiple packages inside
