+++ platform/curator-back/README.md
@@ -1,0 +1,8 @@
+# Developing the backend
+
+## Some guidance
+
+* Backend is written in [Go](https://go.dev/) using [Visual Studio Code](https://code.visualstudio.com/)
+* Start VS Code on this directory
+* working with go modules
+    * all decoupling happens on package level, not module level for this backend implementation
