+++ platform/curator-back/curator-api/.gitignore
@@ -1,0 +1,1 @@
+curator-api
