+++ platform/curator-back/curator-api/api/api.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"net/http"
@@ -6,23 +6,11 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-func main() {
+func StartListen() {
 	router := gin.Default()
 	router.GET("/records", getCuratedRecords)
 
 	router.Run("localhost:8080")
-}
-
-type externalLink struct {
-	ID  string `json:"id"`
-	Url string `json:"url"`
-}
-
-type curatedRecord struct {
-	ID            string         `json:"id"`
-	Headline      string         `json:"headline"`
-	Description   string         `json:"description"`
-	ExternalLinks []externalLink `json:"external-links"`
 }
 
 var records = []curatedRecord{
