--- platform/curator-back/src/data-loader/.gitignore
@@ -1,1 +1,0 @@
-data-loader
