--- platform/curator-back/src/data-loader/main_test.go
@@ -1,12 +1,0 @@
-package main
-
-import (
-	"testing"
-)
-
-func TestInsert(t *testing.T) {
-	// desc := "desc"
-	// head := "head"
-	// record := Insert(desc, head)
-
-}
