+++ platform/setup-go-dev-env.sh
@@ -1,0 +1,8 @@
+ #!/usr/bin/env bash
+
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+
+
+export GOPATH="$SCRIPT_DIR/curator-back"
+
+echo "Set GOPATH env var to $GOPATH"
